website-backup/pyproject.toml
[tool.poetry]
name = "website-backup"
version = "0.1.0"
description = "A project for backing up a website server."
authors = ["Your Name <your.email@example.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^5.4"
boto3 = "^1.18"  # For AWS S3 support, if needed
requests = "^2.26"  # For making HTTP requests, if needed

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = "^21.7b0"
flake8 = "^3.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"